Add a Python 3 NumExpr version, and update all NumPy version to double not float.

+++ Python/pi_python3_numpy_wrong.py
@@ -3,9 +3,9 @@
 #  Calculation of π using quadrature. Sequential algorithm using NumPy – sort of. This is an example of how
 #  not to use NumPy.
 #
-#  Copyright © 2008–2012 Russel Winder
+#  Copyright © 2008–2013  Russel Winder
 
-from numpy import int, float, add, subtract, multiply, divide
+from numpy import int, double, add, subtract, multiply, divide
 from output import out
 from time import time
 
@@ -13,7 +13,7 @@ if __name__ == '__main__':
     n = int(1000000)  # 1000 times fewer than C due to speed issues.
     delta = divide(1.0, n)
     startTime = time()
-    sum = float(0.0)
+    sum = double(0.0)
     for i in range(1, n + 1):
         x = multiply(subtract(i, 0.5), delta)
         sum = add(sum, divide(1.0, add(1.0, multiply(x, x))))
